+++ jhonny/sistema/im/reinicio.php
@@ -1,36 +1,34 @@
-<!DOCTYPE html> <!-- Inicio del documento identificado como HTML5 -->
-<html lang="es">    <!-- Lenguaje español -->
-<head>  <!-- Cabeza de la aplicación web -->
-    <title>PROCESO</title>   <!-- Título de la aplicación web -->
-    <link rel="icon" href="/stp/jhonny/img/favicon.ico" type="image/icon">  <!-- Ícono de la aplicación web -->
-    <meta charset="utf-8">  <!-- Aceptación de carácteres espciales -->
-    <meta name="viewport" content="width=device-width, initial-scale=1">    <!-- Visualización "Responsive" en dispositivos móviles" -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-</head>
+<?php 
 
-<br>
+session_start();
 
-<div class="container">
-<center>
-  <div class="spinner-grow text-danger"></div>
-  <div class="spinner-grow text-danger"></div>
-  <div class="spinner-grow text-danger"></div>
-  <div class="spinner-grow text-danger"></div>
-  <div class="spinner-grow text-danger"></div>
-  <div class="spinner-grow text-danger"></div>
-  <div class="spinner-grow text-danger"></div>
-  <div class="spinner-grow text-danger"></div>
-  <div class="spinner-grow text-danger"></div>
-</center>
-</div>
+if (isset($_SESSION['usuario'])) {
 
-<?php
-  
-  // Se abre la conexión con la DDBB
-  include "../../funciones/conexion.php";
+  switch ($_SESSION['usuario']) {
+
+    case 'constantino':
+    echo '<meta HTTP-EQUIV="REFRESH" CONTENT="0;URL=/stp/">';
+    break;
+
+    case 'diego':
+    echo '<meta HTTP-EQUIV="REFRESH" CONTENT="0;URL=/stp/">';
+    break;
+
+    case 'soledad':
+    echo '<meta HTTP-EQUIV="REFRESH" CONTENT="0;URL=/stp/">';
+    break;
+
+    case 'aleida':
+    echo '<meta HTTP-EQUIV="REFRESH" CONTENT="0;URL=/stp/">';
+    break;
+
+    case 'lily':
+    echo '<meta HTTP-EQUIV="REFRESH" CONTENT="0;URL=/stp/">';
+    break;
+
+    default:
+
+    include "../../funciones/conexion.php";
 
     // Se elimina comletamente todo registro que existe en la tabla, dejándola en un estado de fábrica
     $conexion->query("TRUNCATE TABLE enero_j");
@@ -55,7 +53,43 @@
     // Se cierra la conexión con la DDBB
     include "../../funciones/close.php";
 
+    break;
+  }
+
+} else {
+    echo '<meta HTTP-EQUIV="REFRESH" CONTENT="0;URL=/stp/">';
+  }
+
 ?>
+
+<!DOCTYPE html> <!-- Inicio del documento identificado como HTML5 -->
+<html lang="es">    <!-- Lenguaje español -->
+<head>  <!-- Cabeza de la aplicación web -->
+  <title>PROCESO</title>   <!-- Título de la aplicación web -->
+  <link rel="icon" href="/stp/jhonny/img/favicon.ico" type="image/icon">  <!-- Ícono de la aplicación web -->
+  <meta charset="utf-8">  <!-- Aceptación de carácteres espciales -->
+  <meta name="viewport" content="width=device-width, initial-scale=1">    <!-- Visualización "Responsive" en dispositivos móviles" -->
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
+  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
+</head>
+
+<br>
+
+<div class="container">
+  <center>
+    <div class="spinner-grow text-danger"></div>
+    <div class="spinner-grow text-danger"></div>
+    <div class="spinner-grow text-danger"></div>
+    <div class="spinner-grow text-danger"></div>
+    <div class="spinner-grow text-danger"></div>
+    <div class="spinner-grow text-danger"></div>
+    <div class="spinner-grow text-danger"></div>
+    <div class="spinner-grow text-danger"></div>
+    <div class="spinner-grow text-danger"></div>
+  </center>
+</div>
 
 </body>
 </html>
